feat: Extract hook + add site and depth interval requirements to all screens that use SoilPitInputScreenProps
