+++ dev-client/src/screens/ColorAnalysisScreen/ColorAnalysisScreen.tsx
@@ -17,6 +17,8 @@
 
 import {useMemo, useState} from 'react';
 
+import {useDefaultSiteDepthRequirements} from 'terraso-mobile-client/components/dataRequirements/handleMissingData';
+import {ScreenDataRequirements} from 'terraso-mobile-client/components/dataRequirements/ScreenDataRequirements';
 import {Photo} from 'terraso-mobile-client/components/inputs/image/ImagePicker';
 import {DEFAULT_STACK_NAVIGATOR_OPTIONS} from 'terraso-mobile-client/navigation/constants';
 import {
@@ -46,13 +48,22 @@ export const ColorAnalysisScreen = ({photo, pitProps}: ColorAnalysisProps) => {
     [photo, pitProps, state, setState],
   );
 
+  const requirements = useDefaultSiteDepthRequirements(
+    pitProps.siteId,
+    pitProps.depthInterval.depthInterval,
+  );
+
   return (
-    <ColorAnalysisContext.Provider value={contextValue}>
-      <Stack.Navigator
-        initialRouteName="COLOR_ANALYSIS_HOME"
-        screenOptions={DEFAULT_STACK_NAVIGATOR_OPTIONS}>
-        {screens}
-      </Stack.Navigator>
-    </ColorAnalysisContext.Provider>
+    <ScreenDataRequirements requirements={requirements}>
+      {() => (
+        <ColorAnalysisContext.Provider value={contextValue}>
+          <Stack.Navigator
+            initialRouteName="COLOR_ANALYSIS_HOME"
+            screenOptions={DEFAULT_STACK_NAVIGATOR_OPTIONS}>
+            {screens}
+          </Stack.Navigator>
+        </ColorAnalysisContext.Provider>
+      )}
+    </ScreenDataRequirements>
   );
 };
